requirements.txt
PyYAML
ollama
